+++ apps/humboldt-scoop-cms/android/app/src/main/kotlin/org/kindrobots/humboldt_scoop_cms/MainActivity.kt
@@ -1,0 +1,5 @@
+package org.kindrobots.humboldt_scoop_cms
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()