+++ apps/storymaker/android/app/src/main/kotlin/org/kindrobots/storymaker/MainActivity.kt
@@ -1,0 +1,5 @@
+package org.kindrobots.storymaker
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()