+++ apps/wishmaster/android/app/src/main/kotlin/org/kindrobots/wishmaster/MainActivity.kt
@@ -1,0 +1,5 @@
+package org.kindrobots.wishmaster
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()